Fixed return code 9 through a dirty hack
Pretty nasty stuff, but it made all the test cases worked. I believe the unix detects if the file exists in the /usr/bin/ directory and if it is then it will say it's a valid command. Might want to call a "ls" in the future and parse the filenames into an array and loop through it.